+++ Adroit.Accounting/Adroit.Accounting.Repository/IRepository/IGSTRate.cs
@@ -8,7 +8,7 @@ namespace Adroit.Accounting.Repository.IRepository
         List<GSTRateGridViewModel> List(string connectionString, int loginId = 0, int firmId = 0, string search = "", int pageStart = 0, int pageSize = 10, int sortColumn = 0, string sortOrder = "ASC");
         void Delete(int id, string connectionString);
         List<DropdownViewModel> SelectList(string connectionString);
-        byte Save(GSTRate value, string connectionString);
+        int Save(GSTRate value, string connectionString);
         GSTRateViewModel Get(int id, string connectionString);
     }
 }
